chore(trace): move context module
