+++ ddtrace/opentracer/span_context.py
@@ -4,7 +4,7 @@ from typing import Optional  # noqa:F401
 
 from opentracing import SpanContext as OpenTracingSpanContext
 
-from ddtrace.context import Context as DatadogContext
+from ddtrace._trace.context import Context as DatadogContext
 from ddtrace.internal.compat import NumericType  # noqa:F401
