+++ ddtrace/opentracer/tracer.py
@@ -13,7 +13,7 @@ from opentracing.scope_managers import ThreadLocalScopeManager
 import ddtrace
 from ddtrace import Span as DatadogSpan
 from ddtrace import Tracer as DatadogTracer
-from ddtrace.context import Context as DatadogContext  # noqa:F401
+from ddtrace._trace.context import Context as DatadogContext  # noqa:F401
 from ddtrace.internal.constants import SPAN_API_OPENTRACING
 from ddtrace.internal.utils.config import get_application_name
 from ddtrace.settings import ConfigException
