+++ ddtrace/opentracer/utils.py
@@ -1,6 +1,6 @@
 from opentracing import ScopeManager  # noqa:F401
 
-from ddtrace.provider import BaseContextProvider  # noqa:F401
+from ddtrace._trace.provider import BaseContextProvider  # noqa:F401
 
 
 # DEV: If `asyncio` or `gevent` are unavailable we do not throw an error,
@@ -24,7 +24,7 @@ def get_context_provider_for_scope_manager(scope_manager):
 
         dd_context_provider = ddtrace.contrib.gevent.context_provider
     else:
-        from ddtrace.provider import DefaultContextProvider
+        from ddtrace._trace.provider import DefaultContextProvider
 
         dd_context_provider = DefaultContextProvider()
