+++ ddtrace/tracer.py
@@ -17,6 +17,8 @@ from typing import TypeVar
 from typing import Union
 
 from ddtrace import config
+from ddtrace._trace.context import Context
+from ddtrace._trace.provider import DefaultContextProvider
 from ddtrace.filters import TraceFilter
 from ddtrace.internal.processor.endpoint_call_counter import EndpointCallCounterProcessor
 from ddtrace.internal.sampling import SpanSamplingRule
@@ -30,7 +32,6 @@ from .constants import ENV_KEY
 from .constants import HOSTNAME_KEY
 from .constants import PID
 from .constants import VERSION_KEY
-from .context import Context
 from .internal import agent
 from .internal import atexit
 from .internal import compat
@@ -62,7 +63,6 @@ from .internal.utils.http import verify_url
 from .internal.writer import AgentWriter
 from .internal.writer import LogWriter
 from .internal.writer import TraceWriter
-from .provider import DefaultContextProvider
 from .sampler import BasePrioritySampler
 from .sampler import BaseSampler
 from .sampler import DatadogSampler
