+++ tests/contrib/aiohttp/test_request_safety.py
@@ -7,7 +7,7 @@ import asyncio
 import threading
 from urllib import request
 
-from ddtrace.provider import DefaultContextProvider
+from ddtrace._trace.provider import DefaultContextProvider
 from tests.utils import assert_is_measured
