+++ tests/contrib/gevent/test_tracer.py
@@ -8,7 +8,7 @@ import ddtrace
 from ddtrace.constants import ERROR_MSG
 from ddtrace.constants import SAMPLING_PRIORITY_KEY
 from ddtrace.constants import USER_KEEP
-from ddtrace.context import Context
+from ddtrace._trace.context import Context
 from ddtrace.contrib.gevent import patch
 from ddtrace.contrib.gevent import unpatch
 from tests.opentracer.utils import init_tracer
