+++ tests/tracer/test_context.py
@@ -4,7 +4,7 @@ from typing import Optional  # noqa:F401
 
 import pytest
 
-from ddtrace.context import Context
+from ddtrace._trace.context import Context
 from ddtrace.span import Span
 from ddtrace.tracing._span_link import SpanLink
