+++ tests/tracer/test_correlation_log_context.py
@@ -4,9 +4,9 @@ import structlog
 from ddtrace import Tracer
 from ddtrace import config
 from ddtrace import tracer
-from ddtrace.context import Context
+from ddtrace._trace.context import Context
+from ddtrace._trace.provider import _DD_CONTEXTVAR
 from ddtrace.opentracer.tracer import Tracer as OT_Tracer
-from ddtrace.provider import _DD_CONTEXTVAR
 from tests.utils import override_global_config
